first attempt at nicer email list
