+++ app/services/support/messages/outlook/surface_email_on_case.rb
@@ -4,11 +4,10 @@ module Support
       class SurfaceEmailOnCase
         def self.call(email)
           event_type = "email_#{email.inbox? ? 'from' : 'to'}_school"
-          email_body = email.body.presence || "- No message body -"
 
           email.case.interactions.create!(
             event_type:,
-            body: email_body,
+            body: "Email Interaction",
             additional_data: { email_id: email.id },
             created_at: email.sent_at,
           )
