+++ app/views/support/cases/message_threads/_full_view_message.html.erb
@@ -1,0 +1,14 @@
+<div class="message">
+  <div class="message-details">
+    <span><%= message.sent_by_name %></span>
+    <span><%= message.sent_at_formatted %></span>
+  </div>
+
+  <div class="message-preview govuk-body">
+    <span class="full-view">
+      <%= message.body_for_display(self) %>
+    </span>
+  </div>
+</div>
+
+<hr class="govuk-!-margin-bottom-6" />
