+++ app/views/support/cases/message_threads/_less_than_four_messages.html.erb
@@ -1,0 +1,13 @@
+<% messages.each do |message| %>
+  <div class="message message-condensed">
+    <div class="message-details">
+      <span><%= message.sent_by_name %></span>
+      <span><%= message.sent_at_formatted %></span>
+    </div>
+    <div class="message-preview govuk-body">
+      <span><%= message.body_for_display(self) %></span>
+    </div>
+  </div>
+
+  <hr class="govuk-!-margin-bottom-6" />
+<% end %>
