+++ app/views/support/cases/message_threads/_truncated_message.html.erb
@@ -1,0 +1,18 @@
+<div class="message message-truncated">
+  <div class="message-details">
+    <span><%= message.sent_by_name %></span>
+    <span><%= message.sent_at_formatted %></span>
+  </div>
+
+  <div class="message-preview govuk-body truncated-view" data-component="toggle-truncate">
+    <span class="truncated">
+      <%= message.truncated_body_for_display(self) %>
+    </span>
+
+    <span class="full-view govuk-!-display-none">
+      <%= message.body_for_display(self) %>
+    </span>
+  </div>
+</div>
+
+<hr class="govuk-!-margin-bottom-6" />
